1 规范代码中的Byte和Integer 2 初步完成更新博客的功能

+++ src/main/java/com/seancheer/dao/entity/BaseCategory.java
@@ -26,7 +26,7 @@ public interface BaseCategory {
 
 	String getDescription();
 
-	Integer getId();
+	Byte getId();
 
 	String getName();
